agent: Narrow the responsibilities of RansomwareBitflipEncryptor
